rename frontends to frontend
