+++ frontend/http/writer.go
@@ -18,7 +18,7 @@ import (
 	"net/http"
 
 	"github.com/jzelinskie/trakr/bittorrent"
-	"github.com/jzelinskie/trakr/frontends/http/bencode"
+	"github.com/jzelinskie/trakr/frontend/http/bencode"
 )
 
 // WriteError communicates an error to a BitTorrent client over HTTP.
